refactor : chat list , chat page
