+++ src/hooks/queries/useGetChatRooms.ts
@@ -1,0 +1,10 @@
+import { ChatRoom } from '../../typings';
+import { useQuery } from '@tanstack/react-query';
+import { queryKeys } from '../../shared/constant/queryKeys';
+import { apis } from '../../apis';
+
+function useGetChatRooms() {
+  return useQuery<ChatRoom[]>([queryKeys.CHAT_ROOMS], apis.getChatRooms);
+}
+
+export default useGetChatRooms;
