+++ src/pages/Chat/ChatListPage.tsx
@@ -1,37 +1,34 @@
 import React, { useCallback, useState } from 'react';
 import ChatSearchBar from '../../components/Chat/ChatSearchBar';
 import ChatRoomItem from '../../components/Chat/ChatRoomItem';
-import { useQuery } from '@tanstack/react-query';
-import { apis } from '../../apis';
 import { ChatRoom } from '../../typings';
 import LoadingSpinner from '../../components/Common/loadingSpinner';
 import { useInput } from '../../hooks/util/useInput';
 import { Helmet } from 'react-helmet';
 import { useDebounce } from '../../hooks/util/useDebounce';
+import useGetChatRooms from '../../hooks/queries/useGetChatRooms';
 
 const ChatListPage = () => {
   const { value, handler } = useInput('');
   const [filterData, setFilterData] = useState<ChatRoom[]>([]);
 
-  const { data } = useQuery<ChatRoom[]>(['chatRooms'], apis.getChatRooms, {
-    select: (data) => data.filter((d) => d.title !== value),
-  });
+  const { data: initialRooms } = useGetChatRooms();
 
   const filtering = useCallback(() => {
-    if (!data) {
+    if (!initialRooms) {
       setFilterData([]);
     } else {
       if (value.length === 0) {
-        setFilterData(data);
+        setFilterData(initialRooms);
       }
       const regExp = new RegExp(value, 'g');
-      setFilterData(data.filter((d) => regExp.test(d.title)));
+      setFilterData(initialRooms.filter((d) => regExp.test(d.title)));
     }
-  }, [data, value]);
+  }, [initialRooms, value]);
 
   useDebounce(filtering, 500);
 
-  if (!data) {
+  if (!initialRooms) {
     return <LoadingSpinner />;
   }
